modify expiring / proposed program to use the new ms-numeric-shortcut directive

+++ src/app/business/template/template-business.js
@@ -115,7 +115,7 @@
                 return null;
             }
 
-            return ((parseInt(currentRate) * 100) / parseInt(previousRate)).toFixed(2);
+            return ((parseFloat(currentRate) * 100.0) / parseFloat(previousRate)).toFixed(2);
         }
 
         function getPrintableValue(sectionId)
